refactor(worker): name the alerts state machine by what it decides
state-machine.ts read as the whole worker's, when it only turns a reading
into a severity and decides whether that transition pages. Renamed to
severity-state-machine.ts; imports and the test file follow. No behaviour
change. Addresses the review thread on #1894.
