+++ frontend/worker/src/alerts/scheduler.ts
@@ -18,7 +18,7 @@ import {
   type AlertEvaluationSpec,
 } from "./evaluator-client.js";
 import { logError, logInfo } from "./log.js";
-import { applyAlertStateMachine, deriveAlertSeverity } from "./state-machine.js";
+import { applyAlertStateMachine, deriveAlertSeverity } from "./severity-state-machine.js";
 import { alertWindowStart, computeAlertTick, ALERT_TICK_CRON, type AlertTick } from "./tick.js";
 
 /** One request carries one window pair, so the window token joins the key. */
