+++ frontend/worker/src/alerts/severity-state-machine.ts
